+++ recommender/api/__init__.py
@@ -6,6 +6,7 @@ from recommender.api.schemas import api as models_ns
 from recommender.api.endpoints.recommendations import api as recommendations_ns
 from recommender.api.endpoints.user_actions import api as user_actions_ns
 from recommender.api.endpoints.update import api as update_ns
+from recommender.api.endpoints.health_check import api as health_check_ns
 
 
 api = Api(
@@ -22,3 +23,4 @@ api.add_namespace(models_ns)
 api.add_namespace(recommendations_ns)
 api.add_namespace(user_actions_ns)
 api.add_namespace(update_ns)
+api.add_namespace(health_check_ns)