+++ recommender/services/engine_selector.py
@@ -46,9 +46,11 @@ def get_default_recommendation_alg(
         rec_alg: default recommendation algorithm
     """
     rec_alg = os.environ.get(default_engine, "RL")
-    rec_alg = rec_alg.upper() if len(rec_alg) in {2, 3} else rec_alg.lower()
-
-    return rec_alg if rec_alg in engine_names else "RL"
+    try:
+        index = [e.lower() for e in list(engine_names)].index(rec_alg.lower())
+        return list(engine_names)[index]
+    except ValueError:
+        return "RL"
 
 
 def get_engine_names(